backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiofiles
python-multipart
openai>=1.0.0
python-dotenv
pytest
httpx


